Properly determine if NullString is valid
Right now it looks like this determines if a `NullString` is valid by checking if its `String` is empty when getting the `driver.Value` for writing to a database. This means that if we try try write `{Valid: false, String: "foo"}` to the database, the value won't properly get set to `null`. This is problematic considering the implementation of `UnmarshalJSON` here only sets `Valid = false` and ignores the `String` when unmarshalling `null` from JSON.